Revert "wip"
This reverts commit b8c6399407f77db452f33196106d0cb197c8565a.

+++ archive/src/structures.rs
@@ -72,7 +72,7 @@ impl Meta {
         Some(
             self.canvas_size_changes
                 .iter()
-                .max_by_key(|x| (x.width as u32) * (x.height as u32))?
+                .max_by_key(|x| x.width * x.height)?
                 .clone(),
         )
     }
